Revert "Merge 'Refactor topology out of token_metadata' from Benny Halevy"
This reverts commit 2add9ba2925c732429cd50f71016245ab658e203, reversing
changes made to 7ead1a7857b49ebe9a5c6e66ca5b4640a869da50.
